packages/axip-python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "axip"
version = "0.1.0"
description = "Python SDK for the AXIP agent marketplace protocol"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
keywords = ["axip", "ai", "agents", "marketplace", "protocol"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "websockets>=12.0",
    "PyNaCl>=1.5.0",
]

[project.urls]
Homepage = "https://axiosaiinnovations.com"
Repository = "https://github.com/axiosai/axip"
Documentation = "https://axiosaiinnovations.com/docs"

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
